refactor: moves http files from core to shared. improves expected type for get requests.
