+++ src/app/rundown-view/components/segment-context-menu/segment-context-menu.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 import { SegmentContextMenuComponent } from './segment-context-menu.component'
-import { RundownService } from '../../../core/abstractions/rundown.service'
+import { RundownService } from '../../../shared/abstractions/rundown.service'
 import { instance, mock } from '@typestrong/ts-mockito'
 
 describe('SegmentContextMenuComponent', () => {
