+++ src/app/rundown-view/components/segment/segment.component.spec.ts
@@ -6,7 +6,7 @@ import { instance, mock, when } from '@typestrong/ts-mockito'
 import { Tv2OutputLayerService } from '../../../shared/services/tv2-output-layer.service'
 import { SegmentEntityService } from '../../../core/services/models/segment-entity.service'
 import { PartEntityService } from '../../../core/services/models/part-entity.service'
-import { RundownService } from '../../../core/abstractions/rundown.service'
+import { RundownService } from '../../../shared/abstractions/rundown.service'
 import { SharedModule } from '../../../shared/shared.module'
 import { Logger } from '../../../core/abstractions/logger.service'
 import { TestLoggerFactory } from '../../../test/factories/test-logger.factory'
