+++ src/app/rundown-view/factories/key-binding.factory.ts
@@ -4,7 +4,7 @@ import { StyledKeyBinding } from '../../keyboard/value-objects/styled-key-bindin
 import { ActionService } from '../../shared/abstractions/action.service'
 import { Tv2CameraAction, Tv2RemoteAction, Tv2TransitionAction } from '../../shared/models/tv2-action'
 import { PartActionType } from '../../shared/models/action-type'
-import { RundownService } from '../../core/abstractions/rundown.service'
+import { RundownService } from '../../shared/abstractions/rundown.service'
 import { Rundown } from '../../core/models/rundown'
 import { DialogService } from '../../shared/services/dialog.service'
 import { DialogSeverity } from '../../shared/components/confirmation-dialog/confirmation-dialog.component'
