+++ src/app/shared/abstractions/rundown.service.ts
@@ -1,5 +1,5 @@
 import { Observable } from 'rxjs'
-import { Rundown } from '../models/rundown'
+import { Rundown } from '../../core/models/rundown'
 
 export abstract class RundownService {
   public abstract fetchRundown(rundownId: string): Observable<Rundown>
