+++ src/app/shared/abstractions/show-style-variant.service.ts
@@ -1,5 +1,5 @@
 import { Observable } from 'rxjs'
-import { ShowStyleVariant } from '../models/show-style-variant'
+import { ShowStyleVariant } from '../../core/models/show-style-variant'
 
 export abstract class ShowStyleVariantService {
   public abstract getShowStyleVariant(rundownId: string): Observable<ShowStyleVariant>
