+++ src/app/shared/services/http/http-basic-rundown.service.ts
@@ -2,10 +2,11 @@ import { Injectable } from '@angular/core'
 import { HttpClient } from '@angular/common/http'
 import { HttpErrorService } from './http-error.service'
 import { catchError, map, Observable } from 'rxjs'
-import { BasicRundownService } from '../abstractions/basic-rundown.service'
-import { BasicRundown } from '../models/basic-rundown'
-import { EntityParser } from '../abstractions/entity-parser.service'
-import { environment } from '../../../environments/environment'
+import { BasicRundownService } from '../../abstractions/basic-rundown.service'
+import { BasicRundown } from '../../../core/models/basic-rundown'
+import { EntityParser } from '../../abstractions/entity-parser.service'
+import { environment } from '../../../../environments/environment'
+import { HttpResponse } from './http-response'
 
 @Injectable()
 export class HttpBasicRundownService implements BasicRundownService {
@@ -16,7 +17,7 @@ export class HttpBasicRundownService implements BasicRundownService {
   ) {}
 
   public fetchBasicRundowns(): Observable<BasicRundown[]> {
-    return this.http.get<{ data: unknown }>(`${environment.apiBaseUrl}/rundowns/basic`).pipe(
+    return this.http.get<HttpResponse<BasicRundown[]>>(`${environment.apiBaseUrl}/rundowns/basic`).pipe(
       catchError(error => this.httpErrorService.catchError(error)),
       map(response => this.entityParser.parseBasicRundowns(response.data)) // TODO: Catch this and display/log it
     )
