+++ src/app/shared/services/http/http-response.ts
@@ -1,0 +1,3 @@
+export interface HttpResponse<T> {
+  data: T
+}
