+++ src/app/shared/services/http/http-show-style-variant.service.ts
@@ -2,10 +2,11 @@ import { Injectable } from '@angular/core'
 import { HttpClient } from '@angular/common/http'
 import { catchError, map, Observable } from 'rxjs'
 import { HttpErrorService } from './http-error.service'
-import { EntityParser } from '../abstractions/entity-parser.service'
-import { ShowStyleVariantService } from '../abstractions/show-style-variant.service'
-import { ShowStyleVariant } from '../models/show-style-variant'
-import { environment } from '../../../environments/environment'
+import { EntityParser } from '../../abstractions/entity-parser.service'
+import { ShowStyleVariantService } from '../../abstractions/show-style-variant.service'
+import { ShowStyleVariant } from '../../../core/models/show-style-variant'
+import { environment } from '../../../../environments/environment'
+import { HttpResponse } from './http-response'
 
 const CONFIGURATION_URL: string = `${environment.apiBaseUrl}/configurations`
 
@@ -18,7 +19,7 @@ export class HttpShowStyleVariantService implements ShowStyleVariantService {
   ) {}
 
   public getShowStyleVariant(rundownId: string): Observable<ShowStyleVariant> {
-    return this.http.get<{ data: unknown }>(`${CONFIGURATION_URL}/rundowns/${rundownId}`).pipe(
+    return this.http.get<HttpResponse<ShowStyleVariant>>(`${CONFIGURATION_URL}/rundowns/${rundownId}`).pipe(
       catchError(error => this.httpErrorService.catchError(error)),
       map(response => this.entityParser.parseShowStyleVariant(response.data))
     )
