+++ src/app/shared/services/http/test/http-basic-rundown.service.spec.ts
@@ -1,8 +1,8 @@
-import { HttpBasicRundownService } from './http-basic-rundown.service'
+import { HttpBasicRundownService } from '../http-basic-rundown.service'
 import { HttpClient } from '@angular/common/http'
-import { HttpErrorService } from './http-error.service'
+import { HttpErrorService } from '../http-error.service'
 import { instance, mock } from '@typestrong/ts-mockito'
-import { EntityParser } from '../abstractions/entity-parser.service'
+import { EntityParser } from '../../../abstractions/entity-parser.service'
 
 describe('HttpBasicRundownService', () => {
   it('should be created', () => {
