+++ src/app/shared/services/http/test/http-error.service.spec.ts
@@ -1,8 +1,8 @@
-import { HttpErrorService } from './http-error.service'
+import { HttpErrorService } from '../http-error.service'
 import { MatSnackBar } from '@angular/material/snack-bar'
 import { instance, mock } from '@typestrong/ts-mockito'
-import { TestLoggerFactory } from '../../test/factories/test-logger.factory'
-import { Logger } from '../abstractions/logger.service'
+import { TestLoggerFactory } from '../../../../test/factories/test-logger.factory'
+import { Logger } from '../../../../core/abstractions/logger.service'
 
 describe('HttpErrorService', () => {
   it('should be created', () => {
