+++ src/app/shared/services/http/test/http-show-style-variant.service.spec.ts
@@ -1,8 +1,8 @@
-import { HttpErrorService } from './http-error.service'
+import { HttpErrorService } from '../http-error.service'
 import { instance, mock } from '@typestrong/ts-mockito'
 import { HttpClient } from '@angular/common/http'
-import { EntityParser } from '../abstractions/entity-parser.service'
-import { HttpShowStyleVariantService } from './http-show-style-variant.service'
+import { EntityParser } from '../../../abstractions/entity-parser.service'
+import { HttpShowStyleVariantService } from '../http-show-style-variant.service'
 
 describe('HttpConfigurationService', () => {
   it('should be created', () => {
