+++ src/app/shared/shared.module.ts
@@ -22,7 +22,7 @@ import { FontAwesomeModule } from '@fortawesome/angular-fontawesome'
 import { IconButtonComponent } from './components/icon-button/icon-button.component'
 import { DraggableShelfComponent } from './components/draggable-shelf/draggable-shelf.component'
 import { ActionService } from './abstractions/action.service'
-import { HttpActionService } from './services/http-action.service'
+import { HttpActionService } from './services/http/http-action.service'
 import { ContextMenuComponent } from './components/context-menu/context-menu.component'
 import { CdkMenuModule } from '@angular/cdk/menu'
 import { ActionStateService } from './services/action-state.service'
@@ -35,6 +35,16 @@ import { Tv2ActionPanelComponent } from './components/tv2-action-panel/tv2-actio
 import { Tv2ActionCardComponent } from './components/tv2-action-card/tv2-action-card.component'
 import { TimerPipe } from './pipes/timer/timer.pipe'
 import { MatSelectModule } from '@angular/material/select'
+import { ShowStyleVariantService } from './abstractions/show-style-variant.service'
+import { HttpShowStyleVariantService } from './services/http/http-show-style-variant.service'
+import { RundownService } from './abstractions/rundown.service'
+import { HttpRundownService } from './services/http/http-rundown.service'
+import { BasicRundownService } from './abstractions/basic-rundown.service'
+import { HttpBasicRundownService } from './services/http/http-basic-rundown.service'
+import { HttpClientModule } from '@angular/common/http'
+import { HttpErrorService } from './services/http/http-error.service'
+import { EntityParser } from './abstractions/entity-parser.service'
+import { ZodEntityParser } from '../core/parsers/zod-entity-parser.service'
 
 @NgModule({
   declarations: [
@@ -53,6 +63,7 @@ import { MatSelectModule } from '@angular/material/select'
   ],
   imports: [
     CommonModule,
+    HttpClientModule,
     MatSnackBarModule,
     MatToolbarModule,
     MatIconModule,
@@ -83,8 +94,14 @@ import { MatSelectModule } from '@angular/material/select'
   providers: [
     { provide: MAT_SNACK_BAR_DEFAULT_OPTIONS, useValue: { duration: 5000, verticalPosition: 'top' } },
     { provide: ActionService, useClass: HttpActionService },
+    { provide: ShowStyleVariantService, useClass: HttpShowStyleVariantService },
+    { provide: RundownService, useClass: HttpRundownService },
+    { provide: BasicRundownService, useClass: HttpBasicRundownService },
     { provide: ActionParser, useClass: ZodActionParser },
     { provide: Tv2ActionParser, useClass: ZodTv2ActionParser },
+    { provide: EntityParser, useClass: ZodEntityParser },
+    HttpErrorService,
+    HttpClientModule,
     DialogService,
     ConnectionErrorService,
     Tv2OutputLayerService,
